+++ frontend/src/utils/settings-utils.ts
@@ -1,4 +1,4 @@
-import { Settings } from "#/services/settings";
+import { Settings } from "#/types/settings";
 
 const extractBasicFormData = (formData: FormData) => {
   const provider = formData.get("llm-provider")?.toString();
